[clear] snackbarの記述をまとめた
