+++ app/src/main/kotlin/com/cyder/atsushi/youtubesync/view/activity/SignInActivity.kt
@@ -4,12 +4,10 @@ import android.content.Context
 import android.content.Intent
 import android.databinding.DataBindingUtil
 import android.os.Bundle
-import android.support.design.widget.Snackbar
 import com.cyder.atsushi.youtubesync.R
 import com.cyder.atsushi.youtubesync.databinding.ActivitySignInBinding
-import com.cyder.atsushi.youtubesync.view.helper.hideSoftwareKeyBoard
+import com.cyder.atsushi.youtubesync.view.helper.setUpSnackbar
 import com.cyder.atsushi.youtubesync.viewmodel.SignInActivityViewModel
-import com.cyder.atsushi.youtubesync.viewmodel.SnackbarCallback
 import javax.inject.Inject
 
 
@@ -19,9 +17,9 @@ import javax.inject.Inject
 
 class SignInActivity : BaseActivity() {
 
-    @Inject lateinit var viewModel: SignInActivityViewModel
+    @Inject
+    lateinit var viewModel: SignInActivityViewModel
     private lateinit var binding: ActivitySignInBinding
-    private lateinit var callback: SnackbarCallback
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         getComponent().inject(this)
@@ -30,30 +28,13 @@ class SignInActivity : BaseActivity() {
         binding = DataBindingUtil.setContentView(this, R.layout.activity_sign_in)
         binding.viewModel = viewModel
 
-        setUpSnackbar()
+        viewModel.callback = setUpSnackbar()
     }
 
     override fun onDestroy() {
         viewModel.callback = null
         super.onDestroy()
     }
-
-    private fun setUpSnackbar() {
-        callback = object : SnackbarCallback {
-            override fun onFailed(resId: Int) {
-                currentFocus?.hideSoftwareKeyBoard()
-                Snackbar.make(binding.root,
-                        resId,
-                        Snackbar.LENGTH_SHORT).apply {
-                    setAction(R.string.ok) {
-                        dismiss()
-                    }
-                }.show()
-            }
-        }
-        viewModel.callback = callback
-    }
-
 
 
     companion object {
