+++ app/src/main/kotlin/com/cyder/atsushi/youtubesync/view/activity/SignUpActivity.kt
@@ -4,12 +4,10 @@ import android.content.Context
 import android.content.Intent
 import android.databinding.DataBindingUtil
 import android.os.Bundle
-import android.support.design.widget.Snackbar
 import com.cyder.atsushi.youtubesync.R
 import com.cyder.atsushi.youtubesync.databinding.ActivitySignUpBinding
-import com.cyder.atsushi.youtubesync.view.helper.hideSoftwareKeyBoard
+import com.cyder.atsushi.youtubesync.view.helper.setUpSnackbar
 import com.cyder.atsushi.youtubesync.viewmodel.SignUpActivityViewModel
-import com.cyder.atsushi.youtubesync.viewmodel.SnackbarCallback
 import javax.inject.Inject
 
 /**
@@ -19,7 +17,6 @@ import javax.inject.Inject
 class SignUpActivity : BaseActivity() {
     @Inject lateinit var viewModel: SignUpActivityViewModel
     private lateinit var binding: ActivitySignUpBinding
-    private lateinit var callback: SnackbarCallback
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -29,28 +26,12 @@ class SignUpActivity : BaseActivity() {
         binding = DataBindingUtil.setContentView(this, R.layout.activity_sign_up)
         binding.viewModel = viewModel
 
-        setUpSnackbar()
+        viewModel.callback = setUpSnackbar()
     }
 
     override fun onDestroy() {
         viewModel.callback = null
         super.onDestroy()
-    }
-
-    private fun setUpSnackbar() {
-        callback = object : SnackbarCallback {
-            override fun onFailed(resId: Int) {
-                currentFocus?.hideSoftwareKeyBoard()
-                Snackbar.make(binding.root,
-                        resId,
-                        Snackbar.LENGTH_SHORT).apply {
-                    setAction(R.string.ok) {
-                        dismiss()
-                    }
-                }.show()
-            }
-        }
-        viewModel.callback = callback
     }
 
     companion object {
